% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccboot.r
\name{ccboot}
\alias{ccboot}
\title{Comparacao multipla: Bootstrap}
\usage{
ccboot(
  y,
  trt,
  DFerror,
  SSerror,
  alpha = 0.05,
  group = TRUE,
  main = NULL,
  B = 1000
)
}
\arguments{
\item{y}{Vetor numerico ou complexo contendo a variavel
resposta.}

\item{trt}{Vetor numerico ou complexo contendo os
tratamentos.}

\item{DFerror}{Grau de liberdade do residuo.}

\item{SSerror}{Soma de quadrados do residuo.}

\item{alpha}{Significancia do teste de Bootstrap.}

\item{group}{TRUE ou FALSE}

\item{main}{Titulo}

\item{B}{Numero de reamostragens bootstrap.}
}
\value{
E retornada a comparacao das medias segundo o teste
de Bootstrap.
}
\description{
\code{ccboot} Realiza o teste de Bootstrap de Ramos e
Ferreira (2009) para comparacao multipla de medias.
}
\examples{
data(ex1)
attach(ex1)
dic(trat, ig, quali = TRUE, mcomp='ccboot', sigF = 0.05)
}
\references{
RAMOS, P. S., FERREIRA, D. F. Agrupamento de
medias via bootstrap de populacoes normais e nao-normais,
Revista Ceres, v.56, p.140-149, 2009.
}
\seealso{
\code{\link{snk}}, \code{\link{duncan}},
\code{\link{lsd}}, \code{\link{lsdb}},
\code{\link{scottknott}}, \code{\link{tukey}},
\code{\link{ccf}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Patricia de Siqueira Ramos

Daniel Furtado Ferreira
}
