% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccF.r
\name{ccF}
\alias{ccF}
\title{Teste de comparacoes multiplas de Calinski \& Corsten
baseado na distribuicao F}
\usage{
ccF(y, trt, DFerror, SSerror, alpha = 0.05, group = TRUE, main = NULL)
}
\arguments{
\item{y}{Vetor numerico ou complexo contendo a variavel
resposta.}

\item{trt}{Vetor numerico ou complexo contendo os
tratamentos.}

\item{DFerror}{Grau de liberdade do residuo.}

\item{SSerror}{Soma de quadrados do residuo.}

\item{alpha}{Significancia do teste de Bootstrap.}

\item{group}{TRUE ou FALSE.}

\item{main}{Titulo.}
}
\value{
E retornada a comparacao das medias segundo o teste
de Calinski \& Corsten baseado na distribuicao F.
}
\description{
\code{ccF} Realiza o teste de Calinski \& Corsten baseado
na distribuicao F.
}
\examples{
data(ex2)
attach(ex2)
dbc(trat, provador, aparencia, quali = TRUE, mcomp='ccf',
sigT = 0.05, sigF = 0.05)
}
\references{
CALI\'{N}SKI, T.; CORSTEN, L. C. A. Clustering
means in ANOVA by Simultaneous Testing. Biometrics. v. 41,
p. 39-48, 1985.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Patricia de Siqueira Ramos

Daniel Furtado Ferreira
}
