% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est21Ad.r
\docType{data}
\name{est21Ad}
\alias{est21Ad}
\title{Percevejos no milho: tratamento adicional.}
\format{
Numeric vector.
}
\usage{
data(est21Ad)
}
\description{
Variavel resposta (estatura de plantas de milho)
do tratamento adicional do experimento sobre percevejos.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
