% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex.r
\docType{data}
\name{ex}
\alias{ex}
\title{Videiras: parcelas subdivididas em DBC}
\format{
Um data frame com 24 observacoes das seguintes 4
variaveis:
  \describe{
  \item{\code{trat}}{Fator com os niveis \code{A} e \code{B}}
  \item{\code{dose}}{Vetor numerico}
  \item{\code{rep}}{Vetor numerico}
  \item{\code{resp}}{Vetor numerico}
  }
}
\usage{
data(ex)
}
\description{
Experimento sobre videiras (nao publicado) que
estudou a influencia de diferentes adubos e datas de
colheita no pH das uvas.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
