% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex1.r
\docType{data}
\name{ex1}
\alias{ex1}
\title{Yacon: DIC}
\format{
Um data frame com 24 observacoes das seguintes 2 variaveis.
  \describe{
  \item{\code{trat}}{Vetor numerico.}
  \item{\code{ig}}{Vetor numerico.}
  }
}
\usage{
data(ex1)
}
\description{
Experimento que teve como objetivo avaliar o
consumo da farinha da polpa de yacon sobre o indice glicemico
das dietas experimentais.
}
\references{
RIBEIRO, J. de A. Estudos Quimicos e bioquimicos
do Yacon (Samallanthus sonchifolius) in natura e Processado e
Influencia do seu Consumo sobre Niveis Glicemicos e Lipideos
Fecais de Ratos. 2008. 166p. Dissertation (Master in Food
Science) - Universidade Federal de Lavras, UFLA, Lavras,
2008.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
