% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex2.r
\docType{data}
\name{ex2}
\alias{ex2}
\title{Barras alimenticias: DBC}
\format{
Um data frame com 350 observacoes nas seguintes 3
variaveis.
  \describe{
    \item{\code{provador}}{Vetor numerico}
    \item{\code{trat}}{Fator com os niveis \code{A}
    \code{B} \code{C} \code{D} \code{E}}
    \item{\code{aparencia}}{Vetor numerico}
  }
}
\usage{
data(ex2)
}
\description{
Avaliacao sensorial, em que provadores (blocos)
avaliaram a aparencia das barras alimenticias.
}
\references{
PAIVA, A. P. de. Estudos Tecnologicos, Quimico,
Fisico-quimico e Sensorial de Barras Alimenticias Elaboradas
com Subprodutos e Residuos Agoindustriais. 2008. 131p.
Dissertation (Master in Food Science) - Universidade Federal
de Lavras, UFLA, Lavras, 2008.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
