% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex5.r
\docType{data}
\name{ex5}
\alias{ex5}
\title{Barras alimenticias: fatorial duplo em DBC}
\format{
Um data frame com 160 observacoes nas seguintes 4
variaveis.
  \describe{
    \item{\code{trat}}{Fator com os niveis \code{10g}
    \code{15g} \code{15t} \code{20t}}
    \item{\code{genero}}{Fator com os niveis \code{F}
    \code{M}}
    \item{\code{bloco}}{Vetor numerico.}
    \item{\code{sabor}}{Vetor numerico.}
    }
}
\usage{
data(ex5)
}
\description{
Dados adaptados de uma analise sensorial, onde
provadores de generos diferentes avaliaram o sabor de
diferentes formulacoes de barras alimenticias.
}
\references{
MOREIRA, D. K. T. Extrudados Expandidos de
Arroz, Soja e Gergelim para Uso em Barras Alimenticias.
2010. 166p. Dissertation (Master in Food Science) -
Universidade Federal de Lavras, UFLA, Lavras, 2010.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
