% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex6.r
\docType{data}
\name{ex6}
\alias{ex6}
\title{Dados ficticios 1}
\format{
Um data frame com 24 observacoes nas seguintes 5
variaveis.
  \describe{
   \item{\code{fatorA}}{Vetor numerico.}
   \item{\code{fatorB}}{Vetor numerico.}
   \item{\code{fatorC}}{Vetor numerico.}
   \item{\code{rep}}{Vetor numerico.}
   \item{\code{resp}}{Vetor numerico.}
   }
}
\usage{
data(ex6)
}
\description{
Dados simulados de uma normal com media zero e
variancia 1 para um experimento em esquema de fatorial triplo.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
