% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex7.r
\docType{data}
\name{ex7}
\alias{ex7}
\title{Estatura de plantas de milho 21 dias apos a emergencia.}
\format{
Um data frame com 80 observacoes nas seguintes
4 variaveis.
  \describe{
   \item{\code{periodo}}{Fator com os niveis \code{0-7DAE}
   \code{0-14DAE} \code{0-21DAE} \code{7-14DAE}
   \code{7-21DAE}}
   \item{\code{nivel}}{Vetor numerico.}
   \item{\code{bloco}}{Vetor numerico.}
   \item{\code{est21}}{Vetor numerico.}
   }
   @references RODRIGUES, R. B. Danos do
   percevejo-barriga-verde Dichelops melacanthus (Dallas,
   1851) (Hemiptera: Pentatomidae) na cultura do milho.
   2011. 105f. Dissertacao (Mestrado em Agronomia -
   Universidade Federal de Santa Maria, Santa Maria, 2011.
}
\usage{
data(ex7)
}
\description{
Avaliou-se a estatura de plantas de milho 21
dias apos a emergencia sob infestacao de percevejos
(Dichelops) em diferentes tempos de convivencia (periodo) e
niveis de infestacao (nivel). O tratamento adicional
representa periodo zero e nivel zero.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
