% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex8.r
\docType{data}
\name{ex8}
\alias{ex8}
\title{Compostagem: fatorial duplo com um tratamento adicional em DIC}
\format{
Um data frame com 24 observacoes nas seguintes 5
variaveis.
  \describe{
 \item{\code{inoculante}}{Fator com os niveis \code{esterco}
 e \code{mamona}}.
 \item{\code{biodiesel}}{Vetor numerico.}
 \item{\code{vaso}}{Vetor numerico.}
 \item{\code{fresca}}{Vetor numerico.}
 \item{\code{seca}}{Vetor numerico.}
   }
}
\usage{
data(ex8)
}
\description{
Experimento montado em casas de vegetacao para
observar o desempenho dos compostos obtidos na cultura do
sorgo.
}
\references{
REZENDE, F. A. de. Aproveitamento  da Casca de
Cafe e Borra da Purificacao de Gorduras e Oleos Residuarios
em Compostagem. 2010. 74p. Thesis (Doctorate in
Agronomy/Fitotecny) - Universidade Federal de Lavras, UFLA,
Lavras, 2010.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
