% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex9.r
\docType{data}
\name{ex9}
\alias{ex9}
\title{Coberturas vegetais: parcelas subdivididas em DIC}
\format{
Um data frame com 48 observacoes nas seguintes
4 variaveis.
  \describe{
   \item{\code{cobertura}}{Fator com os niveis \code{T1}
   \code{T2} \code{T3} \code{T4} \code{T5} \code{T6}}
   \item{\code{prof}}{Vetor numerico.}
   \item{\code{rep}}{Vetor numerico.}
   \item{\code{pH}}{Vetor numerico.}
   }
}
\usage{
data(ex9)
}
\description{
Subconjunto de dados de um experimento que
estudou o efeito no pH do solo de coberturas vegetais
submetidas a pisoteio predominantemente por animais bovinos,
sob sistema de pastoreio continuo, analisadas em
profundidades diferentes.
}
\references{
GUERRA, A. R. Atributos de Solo sob Coberturas
Vegetais em Sistema Silvipastoril em Lavras - MG. 2010. 141p.
Dissertation (Master in Forest Engineering) - Universidade
Federal de Lavras, UFLA, Lavras, 2010.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
