% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exnl.r
\docType{data}
\name{exnl}
\alias{exnl}
\title{Exemplo de massa de dados ficticios}
\format{
Um data frame com 8 observacoes das seguintes 8
variaveis.
  \describe{
  \item{\code{trat}}{Vetor numerico.}
  \item{\code{rep}}{Vetor numerico.}
  \item{\code{resp}}{Vetor numerico.}
  }
}
\usage{
data(exnl)
}
\description{
Exemplo de massa de dados ficticios para ajuste
de modelo de regressao nao-linear.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
