% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/han.r
\name{han}
\alias{han}
\title{Teste de Homogeneidade de Variancia de Han}
\usage{
han(resp, trat, bloco)
}
\arguments{
\item{resp}{Vetor numerico ou complexo contendo a variavel
resposta.}

\item{trat}{Vetor numerico ou complexo contendo os
tratamentos.}

\item{bloco}{Vetor numerico ou complexo contendo os blocos.}
}
\value{
Retorna o valor-p do teste de Han de homogeneidade
de variancias para a hipotese nula de que as variancias sao
homogeneas.
}
\description{
\code{han} Realiza o teste de Han (1969) para homogeneidade
de variancias em um delineamento em blocos casualizados.
}
\examples{
data(ex2)
attach(ex2)
dbc(trat, provador, aparencia, hvar = "han")
}
\references{
HAN, C. P. Testing the homogeneity of variances
in a two-way classification. \emph{Biometrics}, 25:153-158,
Mar. 1969.

RIBEIRO, R. \emph{Proposta e comparacao do desempenho de
testes para homogeneidade de variancia de modelos de
classicacao one-way e two-way}. Iniciacao Cientifica.
(Iniciacao Cientifica) - Universidade Federal de Alfenas.
2012.
}
\seealso{
\code{\link{anscombetukey}},
\code{\link{oneilldbc}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @author Marcos Costa de Paula
 @author Mateus Pimenta Siqueira Lima
 @author Renata Ribeiro
}
