% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastC.r
\name{lastC}
\alias{lastC}
\title{Setting the last character of a chain}
\usage{
lastC(x)
}
\arguments{
\item{x}{letters}
}
\value{
x  character.
}
\description{
\code{lastC} A special function for the group of treatments
in the multiple comparison tests. Use order.group.
}
\examples{
x<-c("a","ab","b","c","cd")
lastC(x)
# "a" "b" "b" "c" "d"
}
\seealso{
\code{\link{order.group}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @author Portya Piscitelli Cavalcanti
 (Adapted from Felipe de Mendiburu - GPL)
}
