% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneilldbc.r
\name{oneilldbc}
\alias{oneilldbc}
\title{Teste de Homegeneidade de Variancia: ONeill e Mathews}
\usage{
oneilldbc(resp, trat, bloco)
}
\arguments{
\item{resp}{Vetor numerico ou complexo contendo a variavel}

\item{trat}{Vetor numerico ou complexo contendo os
tratamentos.}

\item{bloco}{Vetor numerico ou complexo contendo os blocos.}
}
\value{
Retorna o valor-p do teste de ONeill e Mathews para
a hipotese nula de que as variancias sao homogeneas.
}
\description{
\code{oneilldbc} Realiza o teste de ONeill e Mathews (2002)
para o teste de homegenidade de variancias para um
delineamento em blocos casualizados.
}
\examples{
data(ex2)
attach(ex2)
dbc(trat, provador, aparencia, hvar = "oneillmathews")
}
\references{
O'NEILL, M. E.; MATHEWS, K. L. Levene tests of
homogeneity of variance for general block and treatment
designs. \emph{Biometrics}, 58:216-224, Mar. 2002.

RIBEIRO, R. \emph{Proposta e comparacao do desempenho de
testes para homogeneidade de variancia de modelos de
classificacao one-way e two-way}. Iniciacao Cientifica.
(Iniciacao Cientifica) - Universidade Federal de Alfenas.
2012.
}
\seealso{
\code{\link{anscombetukey}}, \code{\link{han}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @author Marcos Costa de Paula
 @author Mateus Pimenta Siqueira Lima
}
