% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.group.r
\name{order.group}
\alias{order.group}
\title{Ordering the treatments according to the multiple comparison}
\usage{
order.group(trt, means, N, MSerror, Tprob, std.err, parameter = 1)
}
\arguments{
\item{trt}{Treatments.}

\item{means}{Means of treatment.}

\item{N}{Replications.}

\item{MSerror}{Mean square error.}

\item{Tprob}{Minimum value for the comparison.}

\item{std.err}{Standard error.}

\item{parameter}{Constante 1 (Sd), 0.5 (Sx).}
}
\value{
\itemize{
\item trt Factor
\item means Numeric
\item N Numeric
\item MSerror Numeric
\item Tprob value between 0 and 1
\item std.err Numeric
\item parameter Constant
}
}
\description{
\code{order.group} It orders the groups of means.
}
\seealso{
\code{\link{order.stat.SNK}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @author Portya Piscitelli Cavalcanti
 (Adapted from Felipe de Mendiburu - GPL)
}
