% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.stat.SNK.R
\name{order.stat.SNK}
\alias{order.stat.SNK}
\title{Grouping the treatments averages in a comparison with a
minimum value}
\usage{
order.stat.SNK(treatment, means, minimum)
}
\arguments{
\item{treatment}{Treatment.}

\item{means}{Means of treatment.}

\item{minimum}{Minimum value for the comparison.}
}
\value{
\itemize{
\item trt Factor
\item means Numeric
\item minimum Numeric
}
}
\description{
\code{order.stat.SNK} Orders the groups of means according
to the test of SNK.
}
\seealso{
\code{\link{order.group}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @author Portya Piscitelli Cavalcanti
 (Adapted from Felipe de Mendiburu - GPL)
}
