% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.poly.r
\name{reg.poly}
\alias{reg.poly}
\title{Regressao Polinomial}
\usage{
reg.poly(resp, trat, glres, SQres, gltrat, SQtrat)
}
\arguments{
\item{resp}{Vetor numerico ou complexo contendo a variavel
resposta.}

\item{trat}{Vetor numerico ou complexo contendo os
tratamentos.}

\item{glres}{Grau de liberdade do residuo.}

\item{SQres}{Soma de quadrados do residuo.}

\item{gltrat}{Grau de liberdade do tratamento.}

\item{SQtrat}{Soma de quadrados do tratamento.}
}
\value{
Sao retornados os valores dos ajustes de modelos de
regressao.
}
\description{
\code{reg.poly} Realiza ajustes de modelos sequenciais de
regressao ate o terceiro grau.
}
\references{
GOMES, F. P. Curso de Estatistica Experimental.
10a ed. Piracicaba: ESALQ/USP. 1982. 430.
}
\seealso{
\code{\link{graficos}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
}
