% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respAd.r
\docType{data}
\name{respAd}
\alias{respAd}
\title{Dadaos ficticios: tratamento adicional}
\format{
Vector numerico.
}
\usage{
data(respAd)
}
\description{
Variavel resposta do tratamento adicional.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
