% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samiuddin.r
\name{samiuddin}
\alias{samiuddin}
\title{Teste de homogeneidade de variancias de Samiuddin.}
\usage{
samiuddin(trat, resp, t, r)
}
\arguments{
\item{trat}{Vetor numerico ou complexo contendo os
tratamentos.}

\item{resp}{Vetor numerico ou complexo contendo a
variavel resposta.}

\item{t}{Escalar.Numero de tratamentos.}

\item{r}{Vetor numerico ou complexo contendo o numero de
repeticoes de cada tratamento.}
}
\value{
Retorna o valor-p do teste de homogeneidade de
variancias de Samiuddin para uma hipotese nula de que as
variancias sao homogenenas.
}
\description{
\code{samiuddin} Executa o teste de homogeneidade de
variancias de Samiuddin (1976) para um delineamento
inteiramente casualizado.
}
\examples{
data(ex1)
attach(ex1)
dic(trat, ig, quali = FALSE, hvar = "samiuddin", sigF = 0.05)
}
\references{
SAMIUDDIN, M. Bayesian test of homogeneity of
variance. \emph{Journal of the American Statistical
Association}, 71(354):515-517, Jun. 1976.

NOGUEIRA, D, P.; PEREIRA, G, M. Desempenho de testes para
homogeneidade de variancias em delineamentos inteiramente
casualizados. \emph{Sigmae}, Alfenas, v.2, n.1, p. 7-22.
2013.
}
\seealso{
\code{\link{bartlett}}, \code{\link{layard}},
\code{\link{levene}}, \code{\link{oneillmathews}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @author Marcos Costa de Paula
 @author Mateus Pimenta Siqueira Lima
}
