% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secaAd.r
\docType{data}
\name{secaAd}
\alias{secaAd}
\title{Compostagem: tratamento adicional}
\format{
Vetor numerico.
}
\usage{
data(secaAd)
}
\description{
Variavel resposta (biomassa seca) do tratamento
adicional do experimetno sobre compastagem.
}
\author{
Eric Batista Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
}
\keyword{datasets}
