% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snk.r
\name{snk}
\alias{snk}
\title{Comparacao multipla: SNK}
\usage{
snk(y, trt, DFerror, SSerror, alpha = 0.05, group = TRUE, main = NULL)
}
\arguments{
\item{y}{Vetor numerico ou complexo contendo a variavel
resposta.}

\item{trt}{Vetor numerico ou complexo contendo os
tratamentos.}

\item{DFerror}{Grau de liberdade do residuo.}

\item{SSerror}{Soma de quadrados do residuo.}

\item{alpha}{Significancia do teste de Scott-knott.}

\item{group}{TRUE ou FALSE.}

\item{main}{Titulo.}
}
\value{
E retornada a comparacao das medias segundo o teste
SNK.
}
\description{
\code{snk} Realiza o teste Student Newman Keuls (SNK) para
comparacao multipla de medias.
}
\examples{
data(ex1)
attach(ex1)
dic(trat, ig, quali = TRUE, mcomp = "snk", sigT = 0.05)
}
\seealso{
\code{\link{scottknott}}, \code{\link{duncan}},
\code{\link{lsd}}, \code{\link{lsdb}}, \code{\link{ccboot}},
\code{\link{tukey}}, \code{\link{ccF}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
}
