% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapply.stat.r
\name{tapply.stat}
\alias{tapply.stat}
\title{Statistics of data grouped by factors}
\usage{
tapply.stat(y, x, stat = "mean")
}
\arguments{
\item{y}{Data.frame variables.}

\item{x}{Data.frame factors.}

\item{stat}{Method.}
}
\value{
y Numeric
x Numeric
stat method = "mean", ...
}
\description{
\code{tapply.stat} This process lies in finding statistics
which consist of more than one variable, grouped or crossed
by factors. The table must be organized by columns between
variables and factors.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
(Adapted from Felipe de Mendiburu - GPL)
}
