% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ginv.r
\name{ginv}
\alias{ginv}
\title{Generalized inverse}
\usage{
ginv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{Matriz para a qual deseja-se a inversa de
Moore-Penrose.}

\item{tol}{Uma tolerancia relativa para detectar valores
singulares zero.}
}
\value{
Uma inversa generalizada de Moore-Penrose para X.
}
\description{
\code{ginv} Inversa generalizada
}
\references{
Venables, W. N. and Ripley, B. D. (1999) Modern
Applied Statistics with S-PLUS. Third Edition. Springer.
p.100.
}
\seealso{
\code{\link{solve}}, \code{\link{svd}},
\code{\link{eigen}}
}
