% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotres.r
\name{plotres}
\alias{plotres}
\title{Conjunto de graficos dos residuos do modelo}
\usage{
plotres(a)
}
\arguments{
\item{a}{Objeto contendo a saida da analise do experimento,
tendo esta sido feita utilizando o pacote ExpDes.pt.}
}
\description{
\code{plotres} Graficos dos residuos da saida do modelo.
Conjunto de quatro graficos sao produzidos: (1) Histograma,
(2) Grafico da probabilidade normal dos residuos com bandas
de confianca pela estatistica de ordem, (3) Residuos
Padronizados versus Valores Ajustados e (4) box-plot
(Residuos Padronizados).
}
\examples{
data(ex1)
attach(ex1)
a<-dic(trat, ig)
plotres(a)
}
\references{
STEEL, R. G. D.; TORRIE, J. H. \emph{Principles
and procedures in Statistics: a biometrical approach}.
McGraw-Hill, New York, NY. 1980.
}
\seealso{
\code{\link{graficos}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}
 @author Denismar Alves Nogueira
 @note Esta pode ser utilizada para contrucao dos graficos
 dos residuos de qualquer modelo do ExpDes.pt.
}
