% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.nl.r
\name{reg.nl}
\alias{reg.nl}
\title{Regressao Nao-linear}
\usage{
reg.nl(resp, trat)
}
\arguments{
\item{resp}{Vetor numerico ou complexo contendo a variavel
resposta.}

\item{trat}{Vetor numerico ou complexo contendo os
tratamentos.}
}
\value{
Sao retornados os valores dos ajustes de modelos de
regressao.
}
\description{
\code{reg.nl} Ajusta modelos de regressao nao-linear na ANAVA.
}
\examples{
data(exnl)
attach(exnl)
x<-dic(trat, resp, quali=FALSE, nl=TRUE)
par(mfrow=c(1,2))
graficos(x, grau='pot')
graficos(x, grau='exp')
}
\references{
DRAPER, N.R.; SMITH, H. \emph{Apllied regression
analysis}. 3ed. New York : John Wiley, 1998. 706p.
}
\seealso{
\code{\link{graficos}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Luiz Alberto Beijo
}
