% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scottknott.r
\name{scottknott}
\alias{scottknott}
\title{Comparacao multipla: Scott-Knott}
\usage{
scottknott(y, trt, DFerror, SSerror, alpha = 0.05, group = TRUE, main = NULL)
}
\arguments{
\item{y}{Vetor numerico ou complexo contendo a variavel
resposta.}

\item{trt}{Vetor numerico ou complexo contendo os
tratamentos.}

\item{DFerror}{Grau de liberdade do residuo.}

\item{SSerror}{Soma de quadrados do residuo.}

\item{alpha}{Significancia do teste de Scott-knott.}

\item{group}{TRUE ou FALSE.}

\item{main}{Titulo.}
}
\value{
Retorna uma tabela com as medias dos tratamentos e
seu respectivo agrupamento segundo o teste de Scott-Knott.
}
\description{
\code{scottknott} Aplica o teste Scott-Knott para a
comparacao multipla de medias.
}
\examples{
data(ex1)
attach(ex1)
dic(trat, ig, quali = TRUE, mcomp = "sk", sigT = 0.05)
}
\references{
RAMALHO, M. A. P.; FERREIRA, D. F.; OLIVEIRA,
A. C. de. Experimentacao em Genetica e Melhoramento de
Plantas. 2a ed. Lavras: UFLA. 2005. 300p.
}
\seealso{
\code{\link{snk}}, \code{\link{duncan}},
\code{\link{lsd}}, \code{\link{lsdb}}, \code{\link{ccboot}},
\code{\link{tukey}}, \code{\link{ccF}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
(Adapted from Laercio Junio da Silva - GPL(>=2))
}
