% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.r
\name{tukey}
\alias{tukey}
\title{Comparacao multipla: Tukey}
\usage{
tukey(y, trt, DFerror, SSerror, alpha = 0.05, group = TRUE, main = NULL)
}
\arguments{
\item{y}{Vetor numerico ou complexo contendo a variavel
resposta.}

\item{trt}{Vetor numerico ou complexo contendo os
tratamentos.}

\item{DFerror}{Grau de liberdade do residuo.}

\item{SSerror}{Soma de quadrados do residuo.}

\item{alpha}{Significancia do teste de Scott-knott.}

\item{group}{TRUE ou FALSE.}

\item{main}{Titulo.}
}
\value{
E retornada a comparacao das medias segundo o teste
de Tukey.
}
\description{
\code{tukey} Produz a comparacao multipla de tratamentos
para as medias pela proposta de Tukey. Por padrao considera
a significancia de 0.05.
}
\details{
E necessario produzir a analise de variancia antes.
}
\examples{
data(ex1)
attach(ex1)
dic(trat, ig, quali = TRUE, mcomp = "tukey", sigT = 0.05)
}
\references{
Principles and procedures of statistics a
biometrical approach Steel and Torry and Dickey. Third
Edition 1997
}
\seealso{
\code{\link{scottknott}}, \code{\link{duncan}},
\code{\link{lsd}}, \code{\link{lsdb}}, \code{\link{ccboot}},
\code{\link{snk}}, \code{\link{ccF}}.
}
\author{
Eric B Ferreira,
 \email{eric.ferreira@unifal-mg.edu.br}

Denismar Alves Nogueira

Portya Piscitelli Cavalcanti
(Adapted from Felipe de Mendiburu - GPL)
}
